+++ onnxruntime/contrib_ops/cuda/math/bias_dropout.cc
@@ -1,89 +1,15 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "core/framework/random_seed.h"
-#include "orttraining/training_ops/cuda/nn/dropout.h"
-#include "core/providers/cuda/cuda_common.h"
+#include "contrib_ops/cuda/math/bias_dropout.h"
+#include "core/providers/cuda/nn/dropout.h"
+
 #include "core/providers/common.h"
 
 namespace onnxruntime {
+namespace contrib {
 namespace cuda {
 
-#if defined(CUDA_VERSION) && CUDA_VERSION >= 11000
-#define ALL_IEEE_FLOAT_TENSOR_TYPES {DataTypeImpl::GetTensorType<float>(),      \
-                                     DataTypeImpl::GetTensorType<double>(),     \
-                                     DataTypeImpl::GetTensorType<MLFloat16>(),  \
-                                     DataTypeImpl::GetTensorType<BFloat16>()}
-#define ALL_IEEE_FLOAT_DATA_TYPES float, MLFloat16, double, BFloat16
-#else
-#define ALL_IEEE_FLOAT_TENSOR_TYPES DataTypeImpl::AllIEEEFloatTensorTypes()
-#define ALL_IEEE_FLOAT_DATA_TYPES float, MLFloat16, double
-#endif
-
-#define REGISTER_GRADIENT_KERNEL(OpName)                                 \
-  ONNX_OPERATOR_KERNEL_EX(                                               \
-      OpName,                                                            \
-      kMSDomain,                                                         \
-      1,                                                                 \
-      kCudaExecutionProvider,                                            \
-      KernelDefBuilder()                                                 \
-          .TypeConstraint("T", ALL_IEEE_FLOAT_TENSOR_TYPES)              \
-          .TypeConstraint("T1", ALL_IEEE_FLOAT_TENSOR_TYPES)             \
-          .TypeConstraint("T2", DataTypeImpl::GetTensorType<bool>())     \
-          .InputMemoryType<OrtMemTypeCPUInput>(2),                       \
-      DropoutGrad);
-
-REGISTER_GRADIENT_KERNEL(DropoutGrad)
-
-template <typename T>
-struct DropoutGradComputeImpl {
-  void operator()(cudaStream_t stream,
-                  const int64_t N,
-                  const Tensor& dY,
-                  const bool* mask_data,
-                  const float ratio_data,
-                  Tensor& dX) const {
-    typedef typename ToCudaType<T>::MappedType CudaT;
-
-    const CudaT* dY_data = reinterpret_cast<const CudaT*>(dY.template Data<T>());
-    CudaT* dX_data = reinterpret_cast<CudaT*>(dX.template MutableData<T>());
-    DropoutGradientKernelImpl<CudaT>(stream, N, dY_data, mask_data, ratio_data, dX_data);
-  }
-};
-
-// REVIEW(codemzs): Common out this structure because it is also used in Dropout forward op.
-template <typename T>
-struct GetRatioDataImpl {
-  void operator()(const Tensor* ratio, float& ratio_data) const {
-    ratio_data = static_cast<float>(*(ratio->template Data<T>()));
-    ORT_ENFORCE(ratio_data >= 0.0f && ratio_data < 1.0f, "ratio_data is outside range [0, 1)");
-  }
-};
-
-Status DropoutGrad::ComputeInternal(OpKernelContext* context) const {
-  auto dY = context->Input<Tensor>(0);
-  const TensorShape& shape = dY->Shape();
-  const int64_t N = shape.Size();
-
-  auto mask = context->Input<Tensor>(1);
-  ORT_ENFORCE(mask->Shape().Size() == N);
-  const bool* mask_data = mask->template Data<bool>();
-
-  //Get the ratio_data
-  float ratio_data = default_ratio_;
-  auto ratio = context->Input<Tensor>(2);
-  if (ratio) {
-    utils::MLTypeCallDispatcher<ALL_IEEE_FLOAT_DATA_TYPES> t_disp(ratio->GetElementType());
-    t_disp.Invoke<GetRatioDataImpl>(ratio, ratio_data);
-  }
-
-  auto dX = context->Output(0, shape);
-
-  utils::MLTypeCallDispatcher<ALL_IEEE_FLOAT_DATA_TYPES> t_disp(dY->GetElementType());
-  t_disp.Invoke<DropoutGradComputeImpl>(Stream(), N, *dY, mask_data, ratio_data, *dX);
-
-  return Status::OK();
-}
 
 ONNX_OPERATOR_KERNEL_EX(
     BiasDropout,
@@ -192,4 +118,5 @@ Status BiasDropout::ComputeInternal(OpKernelContext* context) const {
 }
 
 }  // namespace cuda
+}  // namespace contrib
 }  // namespace onnxruntime
