+++ orttraining/orttraining/training_ops/cuda/nn/dropout_grad_impl.h
@@ -17,19 +17,5 @@ void DropoutGradientKernelImpl(
   const float ratio,
   T* dX_data);
 
-template <typename T>
-void BiasDropoutKernelImpl(
-  const cudaDeviceProp& prop,
-  cudaStream_t stream,
-  const int64_t N,
-  const fast_divmod fdm_dim,
-  const float ratio,
-  PhiloxGenerator& generator,
-  const T* X_data,
-  const T* bias_data,
-  const T* residual_data,
-  T* Y_data,
-  bool* mask_data);
-
 }  // namespace cuda
 }  // namespace onnxruntime
